requirements.txt
# Application
fastapi==0.88.0
pydantic==1.10.4
python-dotenv==0.21.0
requests==2.28.1
scrapy==2.7.1
scrapeops-scrapy==0.5.2
scrapeops-scrapy-proxy-sdk==1.0
uvicorn==0.20.0

# Formatters
black==22.12.0
pylint==2.15.9